+++ src/app/(root)/survey/edit/page.tsx
@@ -1,0 +1,7 @@
+import SurveyEdit from '@/components/feature/Survey/Edit';
+
+const page = () => {
+  return <SurveyEdit />;
+};
+
+export default page;
